Lowered tier for local variables in Main, moved all prior functions from main to MT_COMPUTATION.py for array parsing & calculations. Adjusted algorithm to accept an n number of pairs of schedules within schedule.txt and a function in DEBUGGER_TOOL.py to print 10 prewritten test cases. Added two additional algorithm prerequisites to readme file (workers must have 2 existing meetings & MAX 1 contiguous shift in a given day). - These both may be resolved by altering the parsing algorithm from using an odd-even line counter instead of line.startswith("[[")
