+++ schedule.txt
@@ -1,5 +1,50 @@
 [[12:00, 13:00], [16:00, 18:00]]
-[09:00, 19:00]
-[[09:00, 10:30], [12:20, 14:30], [14:30,15:00], [16:00, 17:00 ]]
-[09:00, 18:30]
+[9:00, 19:00]
+[[9:00, 10:30], [12:20, 14:30], [14:30,15:00], [16:00, 17:00 ]]
+[9:00, 18:30]
 30
+[[9:00, 9:30], [16:00, 17:00]]
+[9:00, 17:00]
+[[11:45, 13:30], [13:40, 14:21], [14:30, 14:31], [15:00, 17:00]]
+[11:00, 19:30]
+45
+[[12:00, 9:30], [16:00, 17:00]]
+[12:00, 23:00]
+[[1:45, 3:30], [3:59, 4:01], [4:30, 4:31], [5:00, 5:01]]
+[1:00, 5:30]
+90
+[[2:00, 3:30], [14:00, 18:00], [18:18, 19:00]]
+[00:00, 19:00]
+[[11:05, 11:30], [13:00, 13:21], [13:23, 13:24], [13:27, 13:29]]
+[11:00, 13:30]
+21
+[[5:00, 7:30], [7:31, 7:45]]
+[2:00, 9:00]
+[[11:45, 13:30], [13:40, 14:21], [14:30, 14:31], [15:00, 17:00]]
+[1:11, 21:59]
+55
+[[8:11, 8:49], [14:14, 17:17]]
+[8:00, 20:21]
+[[1:45, 3:40], [3:59, 14:21]]
+[1:12, 14:22]
+10
+[[9:00, 9:30], [16:00, 17:00]]
+[9:00, 17:00]
+[[11:45, 13:30], [15:00, 17:00]]
+[11:00, 19:30]
+39
+[[0:00, 14:30], [14:41, 14:59], [15:01, 15:30], [15:31, 15:45], [16:00, 17:00], [23:00, 23:30]]
+[0:00, 24:00]
+[[11:45, 14:40], [15:30, 17:31], [18:01, 19:19]]
+[0:00, 24:00]
+72
+[[0:00, 0:03], [00:05, 0:18]]
+[0:00, 0:40]
+[[0:15, 0:16], [0:17, 0:21]]
+[0:15, 0:52]
+11
+[[3:10, 5:50], [12:30, 15:56]]
+[2:07, 23:12]
+[[1:16, 1:21], [4:03, 4:41], [5:00, 7:00]]
+[1:08, 10:33]
+99
